User accounts: create user group by default when new users are created
If new users are created, there is now also a user group created with
the same name as the new user's username. This group can be used to
manage edit permissions the user's data. If groups shouldn't be created
automatically, set the NEW_USER_CREATE_USER_GROUP to False in the
settings (behavior so far).

Note that this happens only if users are created through Django, not
when they are created directly in the database.