Tests: assign user group memberships explicitly
This needs to be done, because user groups are now created by default
for each newly created user (with the group name matching the user
name). This prevents us from knowing the group IDs in advance so that we
could specify it in the database fixture. Therefore we need to tadd them
explicitly.