Skeleton API: raise error in importer if 2+ skeletons found for neuron ID
If a neuron ID is passed in, the linked skeleton is deleted. So far an
error has been raised for both zero and 2+ linked skeletons. Now the
error is only raised in the latter case. There is no need to error if
there are zero skeletons. In that case we just create a new one.