Cropping tool: link displayed Z optionally to the bottom of the crop box
A new radio button in front of both the top and the bottom Z index
sliders allows users now to choose whether the link the displayed Z
slice to either the top or the bottom of the copping box. This makes it
easier to visually confirm top and bottom of the bounding box.