Docker: add proper shutdown handler
This allows us to gracefully stop the individual processes that we run
in our image. Without this, Docker will send SIGTERM to PID 1, which is
our entrypoint script at the moment. None of the actual processes
(supervisord, uwsgi, postgres, rabbitmq) hears about the requested stop.
This is changed now so that these processes have a chance to gracefully
shutdown.