Projects: don't let project object manipulate global project reference
The global project object should not be set null (or to anything else
really) from within the project object. Ultimately it would be good to
get rid of the global object, but until then we should at least try to
only update it in init.js to not complicate things.