TrakEM2 import mgmt. command: fix linking of post-synaptic sites
The imported postsynaptic sites where not linked to skeletons imported
from AreaLists, because the wrong ID was stored for imported TrakEM2
objects. Instead of storing the ID of the AreaList, the ID of the XML
data structure was stored, not the object itself. Now, postsynaptic
sites are correctly connected to imported skeletons.