Stack viewer: optionally allow shifting of primary stack
In order to make CATMAID voxel coordinates match Neuroglancer's voxel
coordinates, when a Neuroglancer precomputed dataset with a voxel
offset is used, the primary stack of a stack viewer can now be
optionally shifted by a fixed offset for zoom level zero. This should
match the respective Neuroglancer value and can be defined as metadata
property for a stack.

For example, if the zoom level zero voxel offset for a Neuroglancer
dataset is [-3072, -3072, 0], then the CATMAID stack metadata field can
be edited in the admin view to match this JSON:

{"voxelOffset": [-3072, -3072, 0]}

This is a rather pragmatic way to implement this, and there are more
robust ways of doing this, but given that this is currently a corner
case, this seems to work well enough for now.