Export mgmt. command: fix export of connector tags
So far all tag links for a projevt in the connector_class_instance table
have been exported, if tags shoul be part of the export in the first
place. This is a problem, because first this potentiaally adds a lot of
extra data to the export and, more importantly, second it results in
inconsistent exports: Connectors are referenced by the exported tag
links, without the connector being present. This is fixed now.