Add link widget
This widget allows users to create persistent deep links into the
dataset using an alias. The stored parameters are the same as for
regular deep links, with most of them being optional.

The alias defined for a link can be used to open the deep link, either
by using the link presented in the widget:

<catmaid-url>/<project-id>/links/<alias>

which is only an internal redirect to

<catmaid-url>?pid=<project-id>&link=<alias>

but looks friendlier. Moreover links can be marked private, making them
only accessible (and listable) by the creator of the link.