3D viewer: make volumes loadable from URLs
The widget state is stored in the URL as JSON with single quotes and
volumes are stores as a JSON string as part pf this. For the layout
loading code this doesn't matter, because it uses evil eval. But for
volume loading this is a problem, because it parses this information as
JSON. In order to be able to do that, we need single quotes.

While this is sort of a hack, it only complements the hack to store
volume information as a string. Given this is a special case, it seems
ok.