Layer settings: all stored settings are now restored correctly
Before, the logic to exclude blending and shading in the loop (in order
to handle them separately) wasn't quite right: the else branch was meant
to take care of all other non-special cases, while instead it was
executed only for blending and shading modes.

With this fix now all layer settings can be properly saved as default
and restored again.