Client settings: add extra robustness check to DB storage
In a local test development database I found client settings entries
that would be simply a JSON string rather than an actual object. While
it is not clear when and how this happened, I add an extra test to the
client setting storage method to ensure we put a dict into the database
as settings value and not a string. The front-end was very confused by
the string.