3D viewer: improve robustness of mesh transparency
Transparency with WebGL is hard. The 3D Viewer had trouble in some cases
to display overlapping transparent objects correctly. For instance,
sometimes objects in that were physically behind another object would
appear in front of them.

This fix expands on previous attempts to sort this out: each volume is
still rendered twice, front and back-fragments separately:

1. Transparent front-side fragments
2. Transparent back-side fragments
3. Opaque back-side fragments
4. Opaque front-side fragments

The depth buffer is always updated, except by transparent back-side
fragments and (of course) if fragments are fully transparent.

In order to be able to sort meshes for rendering to respect the physical
ordering, meshes are now transformed to use relative coordinates to
their center of mass and are moved in world space to this center of
mass. With this in place, sorting objects before rendering has been
enabled again and rendering is performed in the correct order.

Fixes catmaid/catmaid#2053