UI: remove location-only link menu entry
This menu entry was redundant with the functionality of the "URL to this
view" link (apart from the fact that it created an alias-link). Since an
alias-link isn't really needed for a location and can be created
explicitly, it seems okay to make the context menu simpler and remove
this entry.