Tracing overlay: fix selecting initial node (e.g. from link) if outside FoV
If an initial node is provided it was so far not selected if it was too
far out of the initial field of view. This is fixed now by making sure
it is loaded separately, if not available already from the viewer.

Now also all stack viewers are updaed.

Thanks to @jasper-tms for reporting this issue!