Tracing tool: remove navigator pointerdown handler correctly
This was supposed to be fixed in edcc4ad26a, and maybe it was at the
time. The added solution doesn't work today anymore though and causes
trouble in some situations. The navigator tool binds its event listeners
to the mouse catcher as well and hence they need to be removed from the
mouse catcher by the tracing tool if it wants to override them.