3D viewer: improve handling of unavailable volumes in widget state
Volumes are now stored with a reference to their project. If this is
found in a local state, the volume is only attempted to load if the
project matches the current one. If no project ID reference is available
on the volume, the volume is attempted to load, but if it fails with a
404 error, the error is silently discarded, assuming the volume was
saved for another project.