Neuron name service: don't call name update handler on new clients
If new clients register skeleton IDs with the neuron name service,
clients were so far called for a potential name update using newly
registered skeletons. This happened so far after the new client was
added to the list of known clients and subsequently the new client was
called as well. This caused unexpected behavior in some widgets, for
instance in the neuron list of the Neuron Navigator one would loose
focus of the filter text input shortly after one started typing right
after the list was opened. This is fixed now by adding the new client
only after other clients are called.

It looks in fact like we don't need to call other clients at all, given
that such an update would only concern otherwise unknown skeletons (and
hence not known to other clients).