API: fix permission URL matching regex
This pattern matched too much, because it didn't require to start at the
beginning of the path. This prevented the admin object permission view
to return the correct response, because the URL suddenly matched our own
permission API (`/permissions`). This is fixed now.

Thanks to @jasper-tms for reporting the issue.

Fixes catmaid/catmaid#2077