Tracing layer: allow deleting nodes "not in view" with Shift + Del
This is mainly useful for deleting nodes in previously visible broken
sections (or nodes otherwise projected onto the current section).

At the moment this will still only allow deletion if the full node state
is already known by the tracing layer. This isn't usually the case if
the node is really out of view. This will lead to an error at the
moment, and can could be added separately.

Fixes catmaid/catmaid#2082