Project listing: make tracing project listing faster (this time really)
In commit 0267305e37e7 I attempted to improve the speed of the project
listing query that will only return projects with actual tracing data
(treenoes or connector or volumes) in them.

The cure was worse than the problem for large datasets with the previous
attempt, but this time the query is actually faster. It uses recursive
queries to implement loose index scans:

https://wiki.postgresql.org/wiki/Loose_indexscan

Unfortunately, we can't just do this on the parent table (location),
because the recursive CTE logic won't be used on the child tables then.
It is good to defined the tracing data conditions explicitly.