Layouts: store only "input models", if available
Some widgets, like the Connectivity Widget, act as a search for
skeletons based on an input set. So far, the layout engine just used all
models provided by a source. In the case of the Connectivity Widget,
this would include all partner skeletons. When such a layout would be
loaded all stored models for a widget would be append to a new widget
instance. For the Connectivity Widget this leads to unintended
consequences: searching for the partners of the whole set, making the
query typically much larger than intended.

This is fixed now by allowing widgets to define a getInputModels()
method, which is supposed to only return those models that can be
considered input of some form to a widget. This is implemented so far
only for the Connectivity Widget and defaults to returning all skeleton
models by default for any other widget.