Connectivity widget: store reference to content DOM element
This is needed for the deferred content creation, because in some cases
it can't assume that the content element is in fact part of the DOM. So
far it looked up the content element in the DOM, which prevented the
widget from being loaded as part of a tab window in the background.

This only becomes an issue if tabs are stored in a layout, like they are
in deep links. Background tabs aren't part of the DOM and therefore
can't be looked up through it. To make the Connectivity Widget load
regardless, a direct reference to the content element is stored now.

This should only been an issue on very few widgets, because most don't
defer the rendering completely like the Connectivity Widget does.