NBLAST: use Postgres' Large Objects for similarity score storage
This migration also adds three new parameter fields to NBLAST
similarities (min length, min soma length, soma tags) and changes the
data type of the result scores. Instead of using a 2D float array
(real[][]), a Postgres large object is used, which is only referenced as
an OID and is stored separately. This makes access slightly more
complex, but is needed in order to store results larger than 1GB. Those
large objects are deleted through tigger functions if the referencing
nblast_similarity entry is updated (and the reference is changed),
deleted or if the table is truncated. The 1GB field size limit of real
is the same as for any other data type. In practice this means a single
NBLAST result score matrix stored in a real[][] field, can contain max
268435456 values, which in turn allows for an all-by-all score matrix of
16384 objects, which isn't all too much. Large objects fix this. They
are also about 25% faster than the previous real[][] version for both
reading and writing score matrices.

This change is done in multiple steps: first the schema changes are
made, but the new scoring column is added with a temporary suffix. Then
the existing scoring is translated and finally the old scoring is
deleted and the new scoring column renamed.

After this change, there will also be no more change history for the
nblast_similarity table. This isn't really needed for this set of
results and would complicate things slightly for this change, because
the large object OID references would need to be tracked as well.

Additionally, the table nblast_similarity_score is added, which will
allow the relational storage of query results. This is mainly useful for
larger result sets and for doing clustering analysis with NBLAST scores.
By default, scores are stored as a large object BLOB, but can optionally
be storaged in a relational fashion. The Neuron Similarity Widget UI has
been updated to include a selection dropd down for the storage mode.

Storing large scoring results (e.g. all by all) usually doesn't need as
much space if stored as a sparse matrix in the form of rows in a table.
It is slightly slower to retrieve as a dense matrix, but this isn't
needed as frequently anyway.