Recommend source build of psycopg2 for production use
According to the release notes of PsycoPg 2.7.4 [1] the psycopg2-binary
package isn't recommended for production use. The problem is that the
referenced SSL versions within the wheels package can in some cases be
different from the system libssl, causing segfaults in psycopg2.
Compiling the module from source works around this.

[1] https://www.psycopg.org/articles/2018/02/08/psycopg-274-released/