Doc: add code of conduct to website docs
There is now a new entry in the "useful links" section on the left hand
side, called "Code of Conduct". This section has also been updated so
that it is rendered slightly separate from the menu above it.

Also, the link to the "Old page" (at MPI CBG) has been removed. The
linked website hasn't been online for a long time and is likely not
coming back.