Layouts: store subscriptions for tracing layers (incl. deep links)
Layouts can generally store subscriptions already, allowing them to be
specified also for deep links. This was so far not possible for
subscriptions to tracing layers (as they aren't widgets like other
common subscription partners). Support for this is added in this commit
by allowing stack viewers to be specified in an extended version within
layouts.

So far stackviewers where specified through the keywords XY, ZY and XZ.
Alternatively, it is now possible to also specify these stack viewers as
{ type: "stackviewer:, orientation: XZ } and similarly for the others.
This extended format also supports a field called "subscriptions", a
list of objects with the  source and target ID of each layer
subscription.

The layout loaded will now prep the stack viewer with a list of default
subscriptions and if a matching layer is loaded later in the process,
the subscription is applied.