Layouts: improve handling of stack viewers in tabbed windows
Stack viewer orientations are currently treated as a special case by the
layout engine: instead of being referenced as full objects with a type
and meta data, they can also be encoded as simple integer numbers to
represent their orientation (e.g. 0 or XY in a layout spec). In layouts,
tabs are represented as MultiNode objects. So far these didn't refer to
stack viewers with their integer orientation. As a consequence, stack
viewers in tab layouts couldn't be loaded and resulted in an error. This
is fixed by this commit.

Fixes catmaid/catmaid#2108