Landmark widget: add default "none" option to target relation setting
A target group relation will instruct the widget to add all those
source-target group relations automatically where the target group is
linked to the source with the selected relation. This is based on a
source group selected by the user.

This isn't implemented for multiple mappings at the moment and the
option shouldn't be visible for this mode. There are likely some use
cases for automatically generating this for many groups, but this will
be implemented separately, if needed.

For now a simple default option "none" will allow multi mappings to be
crated again, if the drop down changed its value.

Fixes catmaid/catmaid#2118