Connectivity widget: remember search term when refreshing partner table
So far the search term wasn't remembered when a partner table was
redrawn. This could be seen for instance when changing the order of the
query neuron table. This is fixed now.

Ideally this would also happen when adding new neurons or removing them
from the query list, and additionally for annotation filters too.

Fixes catmaid/catmaid#2093