3D viewer: set skeleton to invisible if opacity is zero
So far skeletons stayed visible even with an opacity of zero, they were
only rendered with a color of equally zero opacity. This makes them
still have an effect on the depth buffer, which interacts strangely with
how some meshes are rendered (ghosting, dim skeleton shadow). This is
fixed now by not rendering skeletons at all if their opacity is zero.

Thanks to @jasper-tms for reporting this issue!