Neuron navigator: fix annotation listing for subfolder setups
If a subfolder is used (e.g. /catmaid) to host CATMAID, the recent
refactoring of the DataTable setup in the annotation listing broke
loading its data. The reason was that CATMAID.fetch() was called with a
URL generated with CATMAID.makeURL(). The latter however is also called
internally by CATMAID.fetch(), causing CATMAID.makeURL() to applied
twice, leasing a subfolder to be prepended twice as well, resulting in
a non-existent URL. This is fixed by removing the explicit
CATMAID.makeURL() call.

Thanks to @jasper-tms for reporting this issue!