Deep links: improve handling of "URL to this view" Ctrl+Shift+Click
With Ctrl+Shift+Click is used with "URL to this view", a deep link is
created and then opened. Opening this deep link is done with a second
event. The way this was implemented before didn't respect modifiers like
Ctrl or Shift. This is fixed now and Ctrl + Shift will open the newly
created deep link now in a new tab by default, rather than a new window.
This is more consistent with user expectations.

See catmaid/catmaid#2109