Tracing overlay: make line click handling more robust
Skeleton edges have click handlers attached to the pointerupoutside
Pixi.js event. So far this handler tried to focus the stack viewer that
is linked to the node data that is attached to the event object. In
multi-stackviewer environments this doesn't seem to always work,
unfortunately, because the attached node object refers to nodes not
involved in the pointer event. This can cause annoying stackviewer focus
changes in multi-stackviwer layouts.

It's presently not clear why the event handler of a line that wasn't
clicked is called in the first place, but to at least mitigate this
problem by making sure we don't silently change the active stack viewer.