3D Viewer: add option to lock light to camera
This makes it possible to have the light come always from the direction
of the camera, which in turns allows to create the illusion of objects
revolving around the camera target.

The new option can be found as in the Shading tab as a checkbox titled
"Lock light to camera".