Neuron name service: allow caching of name, annotations, meta-ann.
This makes the front-end slightly more intelligent in terms of which
neurons it requests annotations for in an annotationlist update in the
Neuron Name Service. Like before, skeletons unknown to the NNS will
always be requested fully. On subsequent updates, this skeleton becomes
a conditional skeleton request, which will only be fetched and returned
if there was an annotation (and optionally meta-annotation) update since
the last update time. To not send a timestamp for each skeleton (and
make requests at least twice as large), the oldest known update time is
now used as a reference.