Graph widget: don't override edge color state with label color
When updating the edge graphics, e.g. due to a color change, the Graph
Widget so far updated the internal edge color representation of each
edge (edge.data('color')) with the label color. This is however only
needed if the edge coloring mode is set to "source" or "target".
Otherwise it breaks copying the state of the widget (the "wrong"
(overridden) edge color is stored).