Graph widget: reduce number of messages when updating groups on load
When skeletons are appended that are already part of a group, CATMAID
shows a message about the fact that it has been updated within the
group (e.g. color). When adding many skeletons in groups as part of for
instance a deep link, this would result in many messages. To reduce the
noise, a single message is now shown that shows how many skeletons have
been updated.