Deep links: don't require permission to rewrite link URL
For the sake of simpler URLs, deep links can be referred to as
/<project-id/links/<link-id>. The back-end rewrites this then so that
the regular front-end is loaded to ?pid=<project=id>&link=<link-id>.
This required so far browse permission on the project.

Since this particular back-end method simply rewrites a link, no special
permission are needed. This makes it more user-friendly to open a link
while not being logged in. So far, the back-end would simply return an
error. Now the front-end is loaded and more convenient error-handling
can be implemented (like showing a log-in dialog).