Tracing overlay: fix lookup of nearby nodes for Z key
When Z is pressed, one of the first things that are checked is whether
an existing node is close by. The distance to look for other nodes is
defined a constant in pixels (CATCH_RADIUS). This distance has to be
scaled depending on the current zoom level in CATMAID and is expected to
be in stack space by the neighborhood search. Previously the scaling
factor for the CATCH_RADIUS included a multiplication with the lowest
planar resolution of the primary stack. To avoid this, the "scaling"
field is used, rather than the "stackScaling" field.

Additionally, the UI zoom and HiDPI displays are taken into account by
multiplying with window.devicePixelRatio, because on a high DPI screen
or some zoom, the UI generated  coordinate might be based on pixels that
don't actually match what the browser displays.

Fixes catmaid/catmaid#2065