Neuron navigator: remove unneeded filter param in annotation table request
In the current implementation, the annotation list sends a modified
version of the parameter set to the backend that is created by
DataTables itself. This should be changed so that a new object is
created. This commit removed an unneeded search parameter that is part of
the DataTable object. It didn't cause trouble, but caused confusion
during debugging.