Neuron name service: don't notify clients on register() by default
In order to be economic about the data required by the front-end, the
Neuron Name Service keeps track of skeleton IDs and who is interested in
them. If a name representation changes all clients are notified so that
they can update their use of the name.

So far, this client notification was also done after potentially new
skeletons were registered to a client after their names are updated.
This however is counter-intuitive and has lead to problems multiple
times before. It is easy to accidentally cause double updates if
skeletons are registered during the draw/render method of a widget.

The latest instance where this was happening is described in
catmaid/catmaid#2130. There, this double update caused the neuron table
to be rendered twice when loaded initially. Apart from the added
latency, this also made the use of some UI behavior strange, e.g. when
changing the page size for pagination, the table would be re-rendered
due to more/new skeletons fit on one page after a change, causing new
registrations in the Neuron Name Service, which in turn caused the
widget to repaint immediately, overriding the ongoing table expansion.

Therefore, the Neuron Name Service won't notify clients anymore by
default when registering new skeletons. The old behavior can be restored
if the the `notifyClients` parameter is set to `true`.

The notification of clients shouldn't be needed during registration,
because usually the registration is triggered from code that is in the
process of redrawing and updating.

Fixes catmaid/catmaid#2130