Project mgmt. widget: add option to change skeleton IDs
As admin user in a project, it is now possible from to manually change
skeleton IDs from the "ID updates" tab. This is only allowed if the new
ID isn't used already. Also, if the new ID is a new maximum ID number,
all new IDs will be automatically be larger than this ID. Optionally,
the neuron ID can be changed as well.

This also includes a management command named
catmaid_change_skeleton_id, which allows to change a skeleton ID and
optionally a neuron ID as well. It will use the system user to perform
the actual change.

Tests are included, too.