Tests: improve skeleton ID change test
It looks like in some environments the concept ID counter is already
much higher than the new skeleton IDs used for testing the recently
added skeleton ID change functionality. The test so far simply assumed
the new skeleton ID and neuron ID are the current maximum. GitHub
Actions was one place where this assumption broke. Therefore, the new
skeleton ID and neuron ID used in the test are now guaranteed to be
outside of the currently used ID range.