TrakEM2 based node update: fix review preservation
Reviews should only be removed if manually requested through the
review reset parameter. We are modifying the location table and even
though reviews are checked in the treenode (inheritance child) table,
all row-level trigger are executed. Since the review check is a
row-level based trigger, it is executed. This caused the review to be
removed all the time before. To prevent this, we disable the trigger
during the node update.

Thanks to @SanjaJasek for reporting this issue!