TrakEM2 based node update: prevent pending trigger ops during DDL query
As part of this management command, review changing trigger functions are
temporarily disabled. This is done through an ALTER TABLE DDL statement.
To avoid an error that prevents this operation from being undone at the
end of transaction, all triggers need to be executed immediately and
can't be deferred. This at least keeps the implementation simple for
this relatively rarely run command.