TrakEM2 based node update: also disable connector review check trigger
Like with the node change review update trigger, the connector change
review update trigger function needs to be disabled during the node
transformation to have better control over which reviews are changed and
which can remain. So far this wasn't working for nodes that are
synaptic sites and are linked to connector nodes, causing missing
reviews when running a transformation that wasn't supposed to reset
reviews. The problem was the connector review check trigger, which is
now also disabled during the transformation.

Thanks to @SanjaJasek for reporting this problem!