Connectivity table: fix initial selection state logic of added neurons
Skeletons newly added to a Connectivity Table are selected by default,
as indicated by the selected checkbox in the respective top table entry.
This was however only working by accident: newly added skeletons were
not marked as selected by default internally and the logic used to
display a selected or unselected checkbox was backwards. Both are fixed
in this commit.

The problem was detected while looking into catmaid/catmaid#2138.