Layout links: fix right mouse button copying from "URL to this view"
The "URL to this view" link was not updated before, if the Shift key was
pressed after the mouse already moved over the link and the right mouse
button was present. This is fixed now by also updating the link right
before the context menu is shown.

Also, a warning is shown if the link length exceeds 4096 characters,
including the suggestion to also press Ctrl to create a persistent link.