Annotation dialog: add suggested items with escaped comma
Also submit cleaner request by not including an empty annotation in the
list of annotations to be added. This happened as a side effect of
splitting the list of input annotations.

The annotation dialog now also features a preview of the added
annotations. This makes it easier to understand that commas are not
treated as parts of annotations unless escaped.