Docker: update Erlang and RabbitMQ
Ubuntu's RabbitMQ was not recent enough to work with the Erlang 24.x,
which is what was provided by Erlang Solutions, the repository we used
so far. Since we needed a new repository for a more recent RabbitMQ
version, it made sense to also use the Erlang version provided by the
same team (RabbitMQ devs).