3D Viewer: fix export of background color in PNG export
By default PNG exports come with a transparent background. However, if
this is disabled and the displayed background color should be used, the
background color is a white or grey tone. This is fixed with this
commit, by re-rendering the canvas into a new canvas that contains the
correct background color.

Fixes catmaid/catmaid#2161