Client: warn user if Caps-Lock is enabled
The use of Caps-Lock in CATMAID is generally not recommended, key
commands won't work as expected. Users are now warned by default if they
have Caps-Lock enabled and press down a key. This can be disabled from
the Settings Widget (General Settings > Warn if Caps-Lock is enabled).

Fixes catmaid/catmaid#2168