Neuron navigator: fix select-all behavior of neuron lists
The select-all checkboxes in neuron lists, like when showing all neurons
annotated with a particular annotation, so far only selected all neurons
on the currently displayed page. This is now changed so that neurons
that are listed on other pages are now selected as well.

Neuron lists only load the data for the current page and so far the
select-all behavior only looked at this set of neurons. Instead, a
second query is performed now that gets the complete result set (without
scoring, meta-data, etc.).

Fixes catmaid/catmaid#2012
Fixes catmaid/catmaid#2147