Django: continue using 32 bit integers as default ID field
This can be changed separately, if we wanted to. For now the goal is to
get the Django 3.2 upgrade working. In order to update this to BigInt
fields, we would need to update all models that are actually still 32
Bit fields like this:

from django.db import models

class MyModel(models.Model):
    id = models.AutoField(primary_key=True)

Most fields are 64 Bits already, which is just not explicitly reflected
in models.py.