Fix changed return type of Non-ORM JSONB data queries
Since Django 3.1.1, Django started to enforce that all JSONB columns are
returned as strings. This breaks a few queries that now need to explicitly
parse the JSON data. This is what is fixed by this commit.

The upgrade docs only mention this on the side and it would be useful of Django
had made this more clear.

See https://github.com/django/django/commit/438b85dfab4f16a2e709e2bcdbfefecd7bfee89e,
https://github.com/django/django/commit/0be51d2226fce030ac9ca840535a524f41e9832c