Tracing tool: fix straight line distance computation
Straight line distances computed through the measurement button in the
Tracing Tool have been wrong. The API used to get the location of the
involved nodes has a counter-intuitive result: an array of the form [id,
x, y, z], which was accessed by the straight line computation with
indexes 0, 1, 2, leading to wrong coordinate values. This is fixes now
by accessing the correct entries.

Thanks to @wredenc for discovering this bug!

Fixes catmaid/catmaid#2193