Deep links:  fix error on link creation with key shortcut help open
The key shortcut help window doesn't come with an associated widget
object. While this should possibly be fixed too, this commit fixes an
issue with deep link creation where it assumed a widget object to
present in all cases, leading to an error when the key shortcut help is
open.