Landmark widget: improve action buttons for landmark editing
The Edit tab of the Landmark Widget lists individual landmark locations
in a table. Each row contains actions, which so far were hard to read as
separate items. This is fixed now, by using a UI element similar to how
tags are often displayed.

Also, single clicks on the landmark name go to a location now as well.

Fixes catmaid/catmaid#2209