Doc: add numprocs option to supervisor config for Celery
This prevents a threading error to show up in some environments:

django.db.utils.DatabaseError: DatabaseWrapper objects created in a thread can only be used in that same thread. The object with alias 'default' was created in thread id 140089123953704 and this is thread id 140088751748512.

Surprisingly, the default of numprocs should be 1, but setting it
explicitly fixed the error.