WebSockets: make messenging function async
This is needed to make sure there is an event loop available for sending
async websocket messages. Otherwise we would see errors like the
following:

  RuntimeError: Refusing to initialize channel layer without a running event loop.

  If you're writing a django-channels Worker, channels_rabbitmq does not support
  this. Read the way to write workers in the carehare documentation.

  If you're calling `async_to_sync()`, the call must be within code run by
  `sync_to_async()`. Django Channels guarantees this for Django views. Elsewhere,
  beware: `asgiref` *does* let you to call `async_to_sync()` without
  `sync_to_async()`, but `channels_rabbitmq` *doesn't* -- hence this error.

At least with channels_rabbitmq and a CHANNELS_LAYER setting defined,
async messaging seems to fail. This is fixed by marking the message
sending function explicitly as async. This was thankfully suggested in
a GitHub issue of the channels_rabbitmq repository:

https://github.com/CJWorkbench/channels_rabbitmq/issues/34

Unfortunately, this doesn't seem to restore WebSockets messages fully.