WebSockets: fix ASGI message queuing from Celery tasks
This wasn't possible anymore with RabbitMQ since the recent Django and
Channels upgrade, due to the way channels_rabbitmq works. It worked with
Redis though. Luckily, if RabbitMQ is used as broker for both RabbitMQ
and Channels, there is a way around this limitation, as suggested here:

https://github.com/CJWorkbench/channels_rabbitmq/issues/37

This avoids using Channels directly from a Celery task.

See catmaid/catmaid#2204