Landmark widget: ask for confirmation if a transform is added w/o annotation
Not using an annotation can lead to loading many remote skeletons,
which in turn could lock up CATMAID's front-end for a while. To warn
users about this and thus mitigating this danger a bit, a confirmation
dialog is now shown before a transformation is added without an
annotation.

Fixes catmaid/catmaid#2212