Cropping: fix rotation parameter read-out and application
The rotation angle was provided as a POST parameter, but read as a GET
parameter of the HTTP request. This is fixed now.

Also, it turned out right-angle rotations were not applied correctly,
because the image canvas itself was not rotated for those cases, only
the image content. This caused missing content after such rotations.
This is fixed as well. All other angles weren't applied correctly
either, because of using a count-clockwise rotation where a clockwise
one was needed. This looks to me like the Pillow API changed at some
point and we missed this information.