Project management widget: show users w/o any permissions in project
So far the user permission list in the project management widget showed
only users that already had *some* permission for the current project.
Users without any direct permissions were not included. This prevented
an easy workflow where users could register, see a project as read-only
due to automatic group membership, but project admins couldn't set any
other permissions, because they weren't shown in the list of users in
the project management widget.

This is fixed now and all users are shown in the list, as long as they
are visible to the current user. If users are only visible through
project tokens, only those will be shown in this list.

Thanks to Andrew Bellemer for reporting this issue.