Improve Postgres 14 compatibility
Postgres 14 changes the behavior of EXTRACT() to return a numeric type
rather than float8. This leads to extracted epochs to be represented as
decimal.Decimal in Python rather than doubles, which in turn breaks
msgpack encoding with error: "can not serialize 'decimal.Decimal'
object". This is fixed by using the underlying function date_part() as
suggested in the changelog:

https://www.postgresql.org/docs/current/release-14.html