Annotation API: fix query for targets using NOT <annotation>
If annotation targets (e.g. neurons) where looked-up that should NOT
contain a particular annotation, they were so far only returned if the
targets had at least a single annotation. Otherwise they were not
returned.

The reason for this was the behavior of the && operator with a NULL
operand and the boolean evaluation of the resulting NULL result. This is
fixed now by explicitly testing for NULL.

Thanks to @davidhildebrand for reporting this issue.

Fixes catmaid/catmaid#2219